cleanup naming and export similarly to failover controller
