+++ internal/mesh/exports.go
@@ -6,6 +6,7 @@ package mesh
 import (
 	"github.com/hashicorp/consul/internal/controller"
 	"github.com/hashicorp/consul/internal/mesh/internal/controllers"
+	"github.com/hashicorp/consul/internal/mesh/internal/controllers/routes"
 	"github.com/hashicorp/consul/internal/mesh/internal/controllers/sidecarproxy"
 	"github.com/hashicorp/consul/internal/mesh/internal/controllers/sidecarproxy/status"
 	"github.com/hashicorp/consul/internal/mesh/internal/types"
@@ -47,15 +48,16 @@ var (
 
 	// Resource Types for the latest version.
 
-	ProxyConfigurationType     = types.ProxyConfigurationType
-	UpstreamsType              = types.UpstreamsType
-	UpstreamsConfigurationType = types.UpstreamsConfigurationType
-	ProxyStateTemplateType     = types.ProxyStateTemplateType
-	HTTPRouteType              = types.HTTPRouteType
-	GRPCRouteType              = types.GRPCRouteType
-	TCPRouteType               = types.TCPRouteType
-	DestinationPolicyType      = types.DestinationPolicyType
-	ComputedRoutesType         = types.ComputedRoutesType
+	ProxyConfigurationType              = types.ProxyConfigurationType
+	UpstreamsType                       = types.UpstreamsType
+	UpstreamsConfigurationType          = types.UpstreamsConfigurationType
+	ProxyStateTemplateType              = types.ProxyStateTemplateType
+	ProxyStateTemplateConfigurationType = types.ProxyStateTemplateType
+	HTTPRouteType                       = types.HTTPRouteType
+	GRPCRouteType                       = types.GRPCRouteType
+	TCPRouteType                        = types.TCPRouteType
+	DestinationPolicyType               = types.DestinationPolicyType
+	ComputedRoutesType                  = types.ComputedRoutesType
 
 	// Controller statuses.
 
@@ -68,6 +70,19 @@ var (
 	SidecarProxyStatusReasonDestinationServiceFound        = status.StatusReasonDestinationServiceFound
 	SidecarProxyStatusReasonMeshProtocolDestinationPort    = status.StatusReasonMeshProtocolDestinationPort
 	SidecarProxyStatusReasonNonMeshProtocolDestinationPort = status.StatusReasonNonMeshProtocolDestinationPort
+
+	// Routes controller
+	RoutesStatusKey                                                = routes.StatusKey
+	RoutesStatusConditionAccepted                                  = routes.StatusConditionAccepted
+	RoutesStatusConditionAcceptedMissingParentRefReason            = routes.MissingParentRefReason
+	RoutesStatusConditionAcceptedMissingBackendRefReason           = routes.MissingBackendRefReason
+	RoutesStatusConditionAcceptedParentRefOutsideMeshReason        = routes.ParentRefOutsideMeshReason
+	RoutesStatusConditionAcceptedBackendRefOutsideMeshReason       = routes.BackendRefOutsideMeshReason
+	RoutesStatusConditionAcceptedParentRefUsingMeshPortReason      = routes.ParentRefUsingMeshPortReason
+	RoutesStatusConditionAcceptedBackendRefUsingMeshPortReason     = routes.BackendRefUsingMeshPortReason
+	RoutesStatusConditionAcceptedUnknownParentRefPortReason        = routes.UnknownParentRefPortReason
+	RoutesStatusConditionAcceptedUnknownBackendRefPortReason       = routes.UnknownBackendRefPortReason
+	RoutesStatusConditionAcceptedConflictNotBoundToParentRefReason = routes.ConflictNotBoundToParentRefReason
 )
 
 const (
